[release-branch.go1.4] debug/goobj: move to cmd/internal/goobj
??? CL 174250043 / c16349455e05
debug/goobj: move to cmd/internal/goobj

debug/goobj is not ready to be published but it is
needed for the various binary-reading commands.
Move to cmd/internal/goobj.

(The Go 1.3 release branch deleted it, but that's not
an option anymore due to the command dependencies.
The API is still not vetted nor terribly well designed.)

LGTM=adg, dsymonds
R=adg, dsymonds
CC=golang-codereviews
https://codereview.appspot.com/174250043
???

LGTM=rsc
R=rsc
CC=golang-codereviews
https://codereview.appspot.com/177890043